% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{clean_predtweets}
\alias{clean_predtweets}
\title{This function preprocesses prediction files. It uses the non-pred files
  along with some of the other helper functions in this notebook to achieve
  this.}
\usage{
clean_predtweets(file, join_data, colnames = cnames, ...)
}
\arguments{
\item{file}{Path to the prediction file.}

\item{join_data}{A data from of the non-prediction files.}

\item{colnames}{A vector of column names.}

\item{...}{See `transform_prediction`.}
}
\value{
A cleaned tibble for predictions.
}
\description{
This function preprocesses prediction files. It uses the non-pred files
  along with some of the other helper functions in this notebook to achieve
  this.
}
