% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{clean_tweeter}
\alias{clean_tweeter}
\title{Clean up the tweet files upon reading.}
\usage{
clean_tweeter(file)
}
\arguments{
\item{file}{The file to be read. This file must contain a `created_at`,
`from_user`, and `tweet_id_str` column in order to function.}
}
\value{
Cleaned tibble with non_duplicated rows.
}
\description{
This function reads in the file mmanipulates the column types, removes
  duplicaes and creates a couple new variables.
}
