% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{create_comparison}
\alias{create_comparison}
\title{This combines the general and keyword summary data so that comparisons from
This combines the general and keyword summary data so that comparisons from
  baseline can be made.}
\usage{
create_comparison(data, gen_summary_data = gen_ang_summary)
}
\arguments{
\item{data}{The summary data frame.}

\item{gen_summary_data}{The general summary data frame.}
}
\value{
A tibble with an index_diff column.
}
\description{
This combines the general and keyword summary data so that comparisons from
This combines the general and keyword summary data so that comparisons from
  baseline can be made.
}
