% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{create_summary_data}
\alias{create_summary_data}
\title{Summarize the anger index per msa_geoid.}
\usage{
create_summary_data(data)
}
\arguments{
\item{data}{Data frame to be passed.}
}
\value{
Summarized data frame.
}
\description{
Summarize the anger index per msa_geoid.
}
