% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{find_argmax}
\alias{find_argmax}
\title{Find the index with the max argument from a structured set of data.}
\usage{
find_argmax(data)
}
\arguments{
\item{data}{A data frame of the order above.}
}
\value{
A data frame with a new argmax column called `prediction`.
}
\description{
The data columns should be in the following order `tweet_id_str`,
  `job_id`, `created_at`, `none`, `med` and `high`.
}
