% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{pick_msavars}
\alias{pick_msavars}
\title{Pick select msa variables primarily from the `msa_count` data.}
\usage{
pick_msavars(data, var)
}
\arguments{
\item{data}{MSA count data. Most likely `msa_count` data frame.}

\item{var}{MSA variable that you want to compare against the population var.}
}
\value{
A data frame with the selected variable and log of those variables.
}
\description{
This provides easier functionality for quickly selecting variables from data.
}
