% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess the twitter data being by finding the counts for each job.}
\usage{
preprocess(data)
}
\arguments{
\item{data}{The data frame to count the events.}
}
\value{
A data frame with a count column per job_id.
}
\description{
This function just serves a small preprocessing step that will save me from
  writing the same thing over and over and over again. It essentially prunes
  out those who may overtweet more than their analysis alotment, which in
  this case, is set to one.
}
