% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{summarize_emotion}
\alias{summarize_emotion}
\title{Creates a summary of the predictions that make it easier to understand what
  is going on. It does so by providing the number of rows per grouping along
  with a transformation of the prediction variable in which the preiction val
  is divided by two. This creates a mapping from classification to hlaf the
  the size of classification {2 = 1, 1 = 0.5, 0 = 0}.}
\usage{
summarize_emotion(data, summary_var, ...)
}
\arguments{
\item{data}{Dataframe}

\item{summary_var}{A column in `data` to perform the summary on}

\item{...}{The grouping variables.}
}
\description{
Creates a summary of the predictions that make it easier to understand what
  is going on. It does so by providing the number of rows per grouping along
  with a transformation of the prediction variable in which the preiction val
  is divided by two. This creates a mapping from classification to hlaf the
  the size of classification {2 = 1, 1 = 0.5, 0 = 0}.
}
