% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{transform_prediction}
\alias{transform_prediction}
\title{This function maps values to the predictions for the purpose of creating an
  emotional index. Of course, this could be done with the current values but
  with a bit more manipulation on the denominator side of things. This just
  saves a step plus adds a bit more functionality with the manipulation of
  things.}
\usage{
transform_prediction(data, msa_data, col_name, remove_middle = F)
}
\arguments{
\item{data}{The data frame. Must contain a column called `msa_geoid`.}

\item{msa_data}{The msa data frame. Probably from the `haterzmapper` package
and specifically \ref{`hatermapper::topcities`}.}

\item{col_name}{The name of the column to be manipulated.}

\item{remove_middle}{This option removes the middle class if set to true.}
}
\value{
A data frame with cleaned predictions.
}
\description{
This function maps values to the predictions for the purpose of creating an
  emotional index. Of course, this could be done with the current values but
  with a bit more manipulation on the denominator side of things. This just
  saves a step plus adds a bit more functionality with the manipulation of
  things.
}
